JSON Meta Data for the Skirmish Maps
This is a JSON file that contains some basic meta data for the maps.  This is a way that applications can leverage these maps.  One use would be to allow streamers to display the maps or view them within a web site or program from a list of available skirmish maps.